модуль с фильтрами вынес в отдельный модуль
